Move `Semantics`, `CheckedBehavior` and `ModuleKind` to `linker._`.
This is part of the plan for the new linker API #2998.
