+++ test-adapter/src/main/scala/org/scalajs/testadapter/ScalaJSFramework.scala
@@ -13,7 +13,7 @@ import org.scalajs.core.ir.Utils.escapeJS
 
 import org.scalajs.core.tools.io._
 import org.scalajs.core.tools.logging._
-import org.scalajs.core.tools.linker.backend.ModuleKind
+import org.scalajs.core.tools.linker.ModuleKind
 
 import org.scalajs.testadapter.json._
 import org.scalajs.jsenv._
